requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.2
SQLAlchemy>=2.0
psycopg[binary]>=3.1
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
email-validator>=2.1
cloudinary>=1.41
python-dotenv>=1.0
python-multipart